Implementing the exception handling
Implemented some custom exception, right now doing the exception handling, and the reconnect also doing his job.
The exception handling done on the master side.
Also there is a problem if we manually close the slave, then the app doesn't release the connection, and then the master can't automatically rejoin it.